Initial, experimental, support for big (64-bit) integers.
Introduce mfem::bigint type -- long long int.

In class Memory use bigint for sizes, capacity, indices.

In class Table, add support for bigint number of non-zeros. This is
done by dynamically switching between the int array I and the new
bigint array bigI when necessary. Note that the number of rows and
columns in the Table cannot be bigint, only the number of non-zeros.

This extension allows us to handle bigger meshes where Table
objects like the element-to-edge Table can have number of non-zeros
that overflow the int type. For example, a hex mesh with more than
INT_MAX/12+1 (~ 179M) elements overflows the element-to-edge Table.

+++ general/array.hpp
@@ -709,7 +709,7 @@ inline Array<T>::Array(const CT (&values)[N]) : Array(N)
 template <class T>
 inline void Array<T>::GrowSize(int minsize)
 {
-   const int nsize = std::max(minsize, 2 * data.Capacity());
+   const bigint nsize = std::max(bigint(minsize), 2 * data.Capacity());
    Memory<T> p(nsize, data.GetMemoryType());
    p.CopyFrom(data, size);
    p.UseDevice(data.UseDevice());
